refactor: tighten gateway plugin service boundary
